add working path
